adding TuningPairs and TuningProbabilities
